[REFACTOR] :hammer: Refactor cutomer functionality
The customer models fields were throwing errors.
The customer fields were throwing errors because the database had
was not set up to accept null data.

This commit fix this error by setting nulled equals true for
the cutomer fields.
This commit also change the customer models fields and also refactored
the customer app to use django forms. This helps to check for
invalid fields.

This commit also change the updates the customer fields in the
templates.